Change when traceback is printed
Previously, it depended on the verbosity of each specific message. But
I want to be able to turn verbosity up higher for one component and
not get all those tracebacks on the new messages.

So this change now adds tracebacks only if the global verbosity is
turned up to UVM_HIGH or higher.